+++ packages/domain/src/index.ts
@@ -5,6 +5,7 @@
 
 export * from './derive.js';
 export * from './enums.js';
+export * from './errorCodes.js';
 export * from './identity.js';
 export * from './ids.js';
 export * from './schemas.js';